+++ src/cobib/ui/cli.py
@@ -12,11 +12,12 @@ import sys
 from inspect import iscoroutinefunction
 from typing import Any
 
-from rich.console import Console
 from typing_extensions import override
 
 from cobib import __version__
 from cobib.config import config
+from cobib.ui.components import console
+from cobib.ui.shell import Shell
 from cobib.ui.tui import TUI
 from cobib.ui.ui import UI
 from cobib.utils.entry_points import entry_points
@@ -32,15 +33,23 @@ class CLI(UI):
     In addition to the global arguments documented by the base class, the following are supported:
 
       * `--version`: prints the coBib version and quits.
-      * `command`: a single positional argument indicating the name of the command to run. When this
-        is omitted, the `cobib.ui.tui.TUI` gets started.
+      * `-s`, `--shell`: starts the interactive `cobib.ui.shell.Shell`.
+        This is mutually exclusive with executing a sngle `command` (see below).
+      * `command`: a single positional argument indicating the name of the command to run.
+        This is mutually exclusive with starting the interactive `--shell` (see above).
+
+    When **neither** the shell **nor** a command are specified, the `cobib.ui.tui.TUI` gets started.
     """
 
     @override
     def add_extra_parser_arguments(self) -> None:
-        self.parser.add_argument("--version", action="version", version=f"%(prog)s v{__version__}")
+        command_or_shell = self.parser.add_mutually_exclusive_group()
 
-        self.parser.add_argument(
+        command_or_shell.add_argument(
+            "-s", "--shell", action="store_true", help="an interactive shell"
+        )
+
+        command_or_shell.add_argument(
             "command",
             help="the subcommand to be run",
             choices=sorted([cls.name for (cls, _) in entry_points("cobib.commands")]),
@@ -51,6 +60,8 @@ class CLI(UI):
         self.parser.add_argument(
             "args", nargs=argparse.REMAINDER, help="any arguments passed on to the subcommand"
         )
+
+        self.parser.add_argument("--version", action="version", version=f"%(prog)s v{__version__}")
 
     @override
     def parse_args(self) -> argparse.Namespace:
@@ -80,7 +91,7 @@ class CLI(UI):
                 "coBib input arguments.\nIf no subcommand is given, the TUI will be started."
             ),
             epilog=(
-                "Read cobib.1, cobib-commands.7 and cobib-tui.7 for more help.\n"
+                "Read cobib.1, cobib-commands.7, cobib-shell.7 and cobib-tui.7 for more help.\n"
                 "To learn more about the configuration, read cobib-config.5.\n"
                 "If you are new to coBib, cobib-getting-started.7 is a good starting point."
             ),
@@ -90,8 +101,6 @@ class CLI(UI):
         """Runs the CLI interface."""
         arguments = self.parse_args()
 
-        console = Console(theme=config.theme.build())
-
         if not arguments.porcelain:
             # print latest changelog
             changelog = print_changelog(__version__, config.logging.version)
@@ -100,22 +109,34 @@ class CLI(UI):
                 console.print(changelog)  # pragma: no cover
 
         if not arguments.command:
-            # NOTE: we ignore the coverage below because we test the TUI separately
-            task = asyncio.create_task(  # pragma: no cover
-                TUI(verbosity=self._stream_handler.level).run_async()
+            if arguments.porcelain:
+                LOGGER.warning(
+                    "The --porcelain mode has no effect on an interactive UI! Ignoring it..."
+                )
+
+            InteractiveUI = Shell if arguments.shell else TUI
+            task = asyncio.create_task(
+                InteractiveUI(verbosity=self._stream_handler.level).run_async()
             )
-            await task  # pragma: no cover
+            await task
             # the following is required for the asynchronous TUI to quit properly
-            sys.exit()  # pragma: no cover
+            sys.exit()
         else:
-            (cls,) = [
-                cls for (cls, _) in entry_points("cobib.commands") if cls.name == arguments.command
-            ]
-            subcmd = cls.load()(*arguments.args)
+            cmd_cls = self.load_command(arguments.command)
+            if cmd_cls is None:
+                raise RuntimeError(  # pragma: no cover
+                    f"Encountered unexpected error: loading the class for '{arguments.command}' "
+                    "failed! This should not be possible because the argument parser should catch "
+                    "such a case early. When encountering this, please open an issue with steps on "
+                    "how to reproduce this problem: https://gitlab.com/cobib/cobib/-/issues/new"
+                )
+
+            subcmd = cmd_cls(*arguments.args)
             if iscoroutinefunction(subcmd.execute):
                 await subcmd.execute()
             else:
                 subcmd.execute()
+
             if arguments.porcelain:
                 output = subcmd.render_porcelain()
                 for line in output: